requirements.txt
affine==2.4.0
aiobotocore==3.1.2
aioftp==0.21.4
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
bigtree==0.12.5
botocore==1.42.42
bounded-pool-executor==0.0.3
cdsapi==0.7.7
certifi==2025.8.3
cffi==1.17.1
cftime==1.6.5
charset-normalizer==3.4.3
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.3
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
dateparser==1.2.2
DateTime==6.0
dbfread==2.0.7
debugpy==1.8.16
decorator==5.2.1
dotenv==0.9.9
earthaccess==0.16.0
ecmwf-datastores-client==0.4.2
elasticsearch==7.16.2
executing==2.2.1
fastparquet==2024.11.0
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2026.2.0
geobr==0.2.2
geojson==3.2.0
geomet==1.1.0
geopandas==1.1.0
html2text==2025.4.15
html5lib==1.1
humanize==4.13.0
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
jmespath==1.1.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
loguru==0.6.0
lxml==5.4.0
matplotlib==3.10.6
matplotlib-inline==0.1.7
multidict==6.7.1
multimethod==2.0.2
multiurl==0.3.7
nest-asyncio==1.6.0
netCDF4==1.7.3
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pqdm==0.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.21
Pygments==2.19.2
pyhdf==0.11.6
pyogrio==0.12.1
pyparsing==3.2.3
pyproj==3.7.2
pyreaddbc==1.2.0
pysus==1.0.0
python-cmr==0.13.0
python-dateutil==2.8.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
rasterio==1.4.4
regex==2025.9.1
requests==2.32.5
s3fs==2026.2.0
scipy==1.16.3
sentinelsat==1.2.1
shapely==2.1.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
tenacity==9.1.4
tinynetrc==1.3.1
tornado==6.5.2
tqdm==4.64.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==1.26.20
urwid==2.6.16
wcwidth==0.2.13
webencodings==0.5.1
wget==3.2
widgetsnbextension==4.0.14
wrapt==2.1.1
xarray==2025.9.0
yarl==1.22.0
zope.interface==8.1.1
